requirements.txt
Flask
Flask-SQLAlchemy
requests
reportlab
matplotlib
pytest
